+++ src/Spaces/Modifier/SubSpace.jl
@@ -49,7 +49,7 @@ dimension{DS}(sp::SubSpace{DS,Block}) = blocklengths(sp.space)[sp.indexes.K]
 
 
 blocklengths{DS}(sp::SubSpace{DS,UnitRange{Block}}) = blocklengths(sp.space)[Int.(sp.indexes)]
-reindex{DS}(sp::SubSpace{DS,UnitRange{Block}}, ks) = blockstart(sp.space,sp.indexes[1])+ks-1
+reindex{DS}(sp::SubSpace{DS,UnitRange{Block}}, ks)::Int = blockstart(sp.space,sp.indexes[1])+ks-1
 
 dimension{DS}(sp::SubSpace{DS,UnitRange{Block}}) = sum(blocklengths(sp.space)[Int.(sp.indexes)])
